@@ -1,66 +1,74 @@
-# -*- coding: utf-8 -*-
 """Installer for the kitconcept.seo package."""
-
+from pathlib import Path
 from setuptools import find_packages
 from setuptools import setup
 
 
-long_description = "\n\n".join(
-    [
-        open("README.rst").read(),
-        open("CONTRIBUTORS.rst").read(),
-        open("CHANGES.rst").read(),
-    ]
-)
+long_description = f"""
+{Path("README.md").read_text()}\n
+{Path("CONTRIBUTORS.md").read_text()}\n
+{Path("CHANGES.md").read_text()}\n
+"""
 
 
 setup(
     name="kitconcept.seo",
-    version="2.0.2.dev0",
-    description="SEO optimizations plugin for Plone",
+    version="1.0.0a0",
+    description="A new addon for Plone.",
     long_description=long_description,
-    # Get more from https://pypi.python.org/pypi?%3Aaction=list_classifiers
+    long_description_content_type="text/markdown",
     classifiers=[
-        "Development Status :: 5 - Production/Stable",
+        "Development Status :: 3 - Alpha",
         "Environment :: Web Environment",
         "Framework :: Plone",
-        "Framework :: Plone :: 5.2",
+        "Framework :: Plone :: Addon",
+        "Framework :: Plone :: Distribution",
         "Framework :: Plone :: 6.0",
         "Programming Language :: Python",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "Operating System :: OS Independent",
         "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
     ],
-    keywords="Python Plone",
-    author="kitconcept GmbH",
-    author_email="info@kitconcept.com",
-    url="https://pypi.python.org/pypi/kitconcept.seo",
+    keywords="Python Plone CMS",
+    author="Plone Community",
+    author_email="collective@plone.org",
+    url="https://github.com/collective/kitconcept.seo",
+    project_urls={
+        "PyPI": "https://pypi.python.org/pypi/kitconcept.seo",
+        "Source": "https://github.com/collective/kitconcept.seo",
+        "Tracker": "https://github.com/collective/kitconcept.seo/issues",
+    },
     license="GPL version 2",
     packages=find_packages("src", exclude=["ez_setup"]),
     namespace_packages=["kitconcept"],
     package_dir={"": "src"},
     include_package_data=True,
     zip_safe=False,
-    python_requires=">=3.7",
+    python_requires=">=3.8",
     install_requires=[
-        "plone.api",
-        "Products.GenericSetup",
         "setuptools",
-        "z3c.jbot",
-        "plone.behavior",
+        "Plone",
+        "plone.api",
+        "plone.restapi",
     ],
     extras_require={
         "test": [
+            "zest.releaser[recommended]",
+            "zestreleaser.towncrier",
             "plone.app.testing",
-            "plone.testing",
-            "plone.app.contenttypes",
-            "plone.app.robotframework[debug]",
+            "plone.restapi[test]",
+            "pytest",
+            "pytest-cov",
+            "pytest-plone>=0.2.0",
         ],
     },
     entry_points="""
     [z3c.autoinclude.plugin]
     target = plone
+    [console_scripts]
+    update_dist_locale = kitconcept.seo.locales.update:update_locale
     """,
 )
