@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from kitconcept.seo import _
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.dexterity.interfaces import IDexterityContent
@@ -12,7 +11,6 @@ from zope.interface import provider
 
 @provider(IFormFieldProvider)
 class ISeo(model.Schema):
-
     model.fieldset(
         "seo",
         label=_("SEO"),
@@ -118,6 +116,6 @@ class ISeo(model.Schema):
 
 @implementer(ISeo)
 @adapter(IDexterityContent)
-class Seo(object):
+class Seo:
     def __init__(self, context):
         self.context = context
