@@ -1,6 +1,11 @@
-# -*- coding: utf-8 -*-
 """Init and utils."""
-
 from zope.i18nmessageid import MessageFactory
 
-_ = MessageFactory("kitconcept.seo")
+import logging
+
+
+PACKAGE_NAME = "kitconcept.seo"
+
+_ = MessageFactory(PACKAGE_NAME)
+
+logger = logging.getLogger(PACKAGE_NAME)
